+++ tests/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 try:
-    from django.contrib.auth.models import AbstractUser
+    from django.contrib.auth.models import AbstractUser, UserManager
 except ImportError:
     from django.db.models import Model as AbstractUser