+++ tests/test_models.py
@@ -16,6 +16,9 @@ class TestLoginCodes(TestCase):
         self.inactive_user = get_user_model().objects.create(username='inactive', is_active=False)
         self.code = LoginCode.create_code_for_user(self.user)
 
+    def tearDown(self):
+        LoginCode.objects.all().delete()
+
     def test_login_backend(self):
         self.assertEqual(len(self.code.code), 64)
         self.assertIsNotNone(authenticate(username=self.user.username, code=self.code.code))
@@ -38,5 +41,5 @@ class TestLoginCodes(TestCase):
         self.assertIsNone(authenticate(username=self.user.username, code=timeout_code.code))
 
     def test_str(self):
-        code = LoginCode(user=self.user, code='foo', timestamp=datetime(2018, 7, 1))
+        code = LoginCode(user=self.user, timestamp=datetime(2018, 7, 1))
         self.assertEqual(str(code), 'test_user - 2018-07-01 00:00:00')